+++ ee/clickhouse/queries/clickhouse_funnel.py
@@ -4,28 +4,39 @@ from django.utils import timezone
 
 from ee.clickhouse.client import sync_execute
 from ee.clickhouse.models.property import parse_prop_clauses
+from ee.clickhouse.queries.clickhouse_funnel_base import ClickhouseFunnelBase
 from ee.clickhouse.queries.clickhouse_funnel_trends import ClickhouseFunnelTrends
 from ee.clickhouse.queries.util import parse_timestamps
 from ee.clickhouse.sql.funnels.funnel import FUNNEL_SQL
 from ee.clickhouse.sql.person import GET_LATEST_PERSON_DISTINCT_ID_SQL
 from posthog.constants import TRENDS_LINEAR
-from posthog.models.filters import Filter
-from posthog.models.team import Team
-from posthog.queries.funnel import Funnel
 from posthog.utils import relative_date_parse
 
 
-class ClickhouseFunnel(Funnel):
-    _filter: Filter
-    _team: Team
+class ClickhouseFunnel(ClickhouseFunnelBase):
+    def run(self, *args, **kwargs) -> List[Dict[str, Any]]:
+        if len(self._filter.entities) == 0:
+            return []
 
-    def __init__(self, filter: Filter, team: Team) -> None:
-        self._filter = filter
-        self._team = team
-        self.params = {
-            "team_id": self._team.pk,
-            "events": [],  # purely a speed optimization, don't need this for filtering
-        }
+        if self._filter.display == TRENDS_LINEAR:
+            return ClickhouseFunnelTrends(self._filter, self._team).run()
+        else:
+            # Format of this is [step order, person count (that reached that step), array of person uuids]
+            results = self._exec_query()
+
+            steps = []
+            relevant_people = []
+            total_people = 0
+
+            for step in reversed(self._filter.entities):
+                # Clickhouse step order starts at one, hence the +1
+                result_step = [x for x in results if step.order + 1 == x[0]]
+                if len(result_step) > 0:
+                    total_people += result_step[0][1]
+                    relevant_people += result_step[0][2]
+                steps.append(self._serialize_step(step, total_people, relevant_people[0:100]))
+
+            return steps[::-1]  #  reverse
 
     def _exec_query(self) -> List[Tuple]:
         prop_filters, prop_filter_params = parse_prop_clauses(
@@ -62,27 +73,3 @@ class ClickhouseFunnel(Funnel):
             latest_distinct_id_sql=GET_LATEST_PERSON_DISTINCT_ID_SQL,
         )
         return sync_execute(query, self.params)
-
-    def run(self, *args, **kwargs) -> List[Dict[str, Any]]:
-        if len(self._filter.entities) == 0:
-            return []
-
-        if self._filter.display == TRENDS_LINEAR:
-            return ClickhouseFunnelTrends(self._filter, self._team).run()
-        else:
-            # Format of this is [step order, person count (that reached that step), array of person uuids]
-            results = self._exec_query()
-
-            steps = []
-            relevant_people = []
-            total_people = 0
-
-            for step in reversed(self._filter.entities):
-                # Clickhouse step order starts at one, hence the +1
-                result_step = [x for x in results if step.order + 1 == x[0]]
-                if len(result_step) > 0:
-                    total_people += result_step[0][1]
-                    relevant_people += result_step[0][2]
-                steps.append(self._serialize_step(step, total_people, relevant_people[0:100]))
-
-            return steps[::-1]  #  reverse
